// types.ts
import { Timestamp } from './src/utils/firebase';

export interface Car {
  condition: string;
  carID: string;
  userID: string;
  userPhotoUrl: string;
  vin: string;
  milage: number;
  dayOfFirstRegistration: number;
  monthOfFirstRegistration: number;
  yearOfFirstRegistration: number;
  generation: string;
  make: string;
  model: string;
  fuelType: string;
  madeYear: number;
  power: number;
  capacity: number;
  DoorNumber: number;
  transmission: string;
  driveTrain: string;
  color: string;
  price: number;
  shortDescription: string;
  images: string[];
  tags: string[];
  version: string;
  EmmisionStandard: string;
  bodyWork: string;
  youtubeLink: string;
  longDescription?: string;
  countryOfOrigin: string;
  warrantyKm: number;
  country: string;
  city: string;
  createdAt: Timestamp;
  c02Emission: number;
  searchKeywords: string[];
}


export type CarDocument = {
  tags: string[];
  make: string;
  color: string;
  capacity: number;
  country: string;
  dayOfFirstRegistration: number;
  power: number;
  countryOfOrigin: string;
  madeYear: number;
  version: string;
  monthOfFirstRegistration: number;
  c02Emission: number;
  price: number;
  userID: string;
  condition: string;
  shortDescription: string;
  driveTrain: string;
  searchKeywords: string[];
  model: string;
  DoorNumber: number;
  vin: string;
  images: string[];
  userPhotoUrl: string;
  carID: string;
  youtubeLink: string;
  transmission: string;
  generation: string;
  fuelType: string;
  longDescription: string;
  createdAt: Timestamp;
  city: string;
  milage: number;
  bodyWork: string;
  yearOfFirstRegistration: number;
  warrantyKm: number;
  EmmisionStandard: string;
  id: string;
  userPhoto: string;
};